+++ src/plone.server/plone/server/catalog/index.py
@@ -1,12 +1,11 @@
 # -*- coding: utf-8 -*-
-from plone.server.catalog.interfaces import ICatalogUtility
+from plone.server.interfaces import ICatalogUtility
+from plone.server.interfaces import ISite
 from plone.server.transactions import get_current_request
 from plone.server.transactions import RequestNotFound
 from plone.server.transactions import tm
-from zope.component import getUtility
 from zope.component import queryUtility
 
-import asyncio
 import logging
 import transaction
 
@@ -16,19 +15,20 @@ logger = logging.getLogger('plone.server')
 
 class CommitHook(object):
 
-    def __init__(self, site_id, loop):
+    def __init__(self, site, request):
         self.remove = []
         self.index = {}
-        self.site_id = site_id
-        self.loop = loop
+        self.site = site
+        self.request = request
 
     async def __call__(self, trns):
         if not trns:
             return
         # Commits are run in sync thread so there is no asyncloop
-        search = getUtility(ICatalogUtility)
-        await search.remove(self.remove, self.site_id)
-        await search.index(self.index, self.site_id)
+        search = queryUtility(ICatalogUtility)
+        if search:
+            await search.remove(self.site, self.remove)
+            await search.index(self.site, self.index)
 
         self.index = {}
         self.remove = []
@@ -36,11 +36,16 @@ class CommitHook(object):
 
 def get_hook():
 
-    search = queryUtility(ICatalogUtility)
+    request = get_current_request()
+    try:
+        site = request.site
+        search = queryUtility(ICatalogUtility)
+    except (AttributeError, KeyError):
+        return
+
     if not search:
         return  # no search configured
 
-    request = get_current_request()
     try:
         trns = tm(request).get()
     except RequestNotFound:
@@ -51,21 +56,21 @@ def get_hook():
             hook = _hook[0]
             break
     if hook is None:
-        loop = asyncio.get_event_loop()
-        hook = CommitHook(request._site_id, loop)
+        hook = CommitHook(site, request)
         trns.addAfterCommitHook(hook)
     return hook
 
 
 def remove_object(obj, event):
-    hook = get_hook()
-    if hook is None:
-        return
     uid = getattr(obj, 'uuid', None)
     if uid is None:
         return
     portal_type = getattr(obj, 'portal_type', None)
-    if portal_type is None:
+    if portal_type is None or ISite.providedBy(obj):
+        return
+
+    hook = get_hook()
+    if hook is None:
         return
     hook.remove.append((uid, portal_type))
     if uid in hook.index:
@@ -73,27 +78,28 @@ def remove_object(obj, event):
 
 
 def add_object(obj, event):
-    hook = get_hook()
-    if hook is None:
-        return
     uid = getattr(obj, 'uuid', None)
     if uid is None:
         return
     portal_type = getattr(obj, 'portal_type', None)
-    if portal_type is None:
+    if portal_type is None or ISite.providedBy(obj):
+        return
+
+    hook = get_hook()
+    if hook is None:
         return
     search = queryUtility(ICatalogUtility)
-    if search is not None:
+    if search:
         hook.index[uid] = search.get_data(obj)
 
 
-async def add_index(obj, event):
+async def initialize_catalog(site, event):
     search = queryUtility(ICatalogUtility)
-    if search is not None:
-        await search.create_index(obj.id)
+    if search:
+        await search.initialize_catalog(site)
 
 
-async def remove_index(obj, event):
+async def remove_catalog(site, event):
     search = queryUtility(ICatalogUtility)
-    if search is not None:
-        await search.remove_index(obj.id)
+    if search:
+        await search.remove_catalog(site)