+++ src/plone.server/plone/server/tests/test_catalog.py
@@ -1,0 +1,35 @@
+# -*- encoding: utf-8 -*-
+from plone.server.testing import PloneServerBaseTestCase
+from plone.server.catalog.utils import get_index_fields
+from plone.server.content import createContentInContainer, createContent
+from plone.server.interfaces import ICatalogDataAdapter
+
+
+class TestCatalog(PloneServerBaseTestCase):
+    """Functional testing of the API REST."""
+
+    def test_indexed_fields(self):
+        fields = get_index_fields('Item')
+        self.assertTrue('portal_type' in fields)
+        self.assertTrue('uuid' in fields)
+        self.assertTrue('path' in fields)
+        self.assertTrue('title' in fields)
+        self.assertTrue('creation_date' in fields)
+
+    def test_get_index_data(self):
+        self.login()
+        db = self.layer.app['plone']
+        site = createContent(
+            'Site',
+            id='plone',
+            title='Plone')
+        site.__name__ = 'plone'
+        db['plone'] = site
+        ob = createContentInContainer(site, 'Item', 'foobar')
+
+        data = ICatalogDataAdapter(ob)
+        fields = data()
+        self.assertTrue('portal_type' in fields)
+        self.assertTrue('uuid' in fields)
+        self.assertTrue('path' in fields)
+        self.assertTrue('title' in fields)